Enable typescript and organize gatsby node files
