+++ .vscode/settings.json
@@ -5,22 +5,8 @@
 
   "editor.wordWrap": "on",
 
-  "files.associations": {
-    "*.js": "javascriptreact"
-  },
-
-  "liquid.format": false,
-
-  "[ruby]": {
-    "editor.detectIndentation": false,
-    "editor.tabSize": 2,
-    "editor.formatOnSave": true
-  },
-
-  "[html]": {
-    "editor.detectIndentation": false,
-    "editor.tabSize": 2,
-    "editor.formatOnSave": false
-  },
-  "search.usePCRE2": true
+  "typescript.tsdk": "node_modules/typescript/lib",
+  "editor.codeActionsOnSave": {
+    "source.fixAll.eslint": true
+  }
 }
