+++ package.json
@@ -18,6 +18,10 @@
     "@sentry/gatsby": "^5.20.1",
     "@sentry/react": "^5.20.0",
     "@sentry/tracing": "^5.20.1",
+    "@types/node": "^12.18.3",
+    "@types/react": "^16.9.46",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-helmet": "^6.1.0",
     "add": "^2.0.6",
     "algoliasearch": "^4.2.0",
     "babel-preset-gatsby": "^0.4.1",
@@ -36,6 +40,7 @@
     "gatsby-plugin-react-helmet": "^3.3.10",
     "gatsby-plugin-sass": "^2.3.12",
     "gatsby-plugin-sharp": "^2.6.24",
+    "gatsby-plugin-typescript": "^2.4.18",
     "gatsby-plugin-zeit-now": "^0.3.0",
     "gatsby-remark-autolink-headers": "^2.3.1",
     "gatsby-remark-check-links": "^2.1.0",
@@ -57,12 +62,15 @@
     "react-feather": "^2.0.8",
     "react-helmet": "^5.2.1",
     "react-instantsearch-dom": "^6.4.0",
-    "remark-deflist": "^0.2.1"
+    "remark-deflist": "^0.2.1",
+    "ts-node": "^8.10.2",
+    "typescript": "^3.9.7"
   },
   "scripts": {
     "build": "GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES=true gatsby build --log-pages",
     "sdk:copy": "node scripts/sdk-symlink",
     "develop": "gatsby develop -p 3000",
+    "lint": "tsc --skipLibCheck",
     "start": "npm run develop",
     "format": "prettier --write \"src/**/*.js\"",
     "test": "GATSBY_ENV=test gatsby build"
