+++ src/gatsby/onCreateWebpackConfig.ts
@@ -1,0 +1,11 @@
+import path from "path";
+
+export default ({ _stage, actions }) => {
+  actions.setWebpackConfig({
+    resolve: {
+      alias: {
+        "~src": path.join(path.resolve(__dirname, "..")),
+      },
+    },
+  });
+};
